+++ src/components/Header.tsx
@@ -5,7 +5,9 @@ export default function Header() {
 
   return (
     <div className="h-16 bg-gray-800 border-b content-center text-center">
-      <span className="font-bold text-xl text-white">{t('appName')}</span>
+      <span className="font-bold text-xl text-white">
+        {t('common.appName')}
+      </span>
     </div>
   )
 }