+++ src/models/lunaticType.ts
@@ -1,6 +1,7 @@
 import {
-  type LunaticOptions, type LunaticSource, getArticulationState,
-  useLunatic,
+  type LunaticOptions,
+  type LunaticSource,
+  type LunaticState,
 } from '@inseefr/lunatic'
 
 export type LunaticGetReferentiel = LunaticOptions['getReferentiel']
@@ -9,13 +10,13 @@ export type Nomenclature = Awaited<
   ReturnType<NonNullable<LunaticGetReferentiel>>
 >
 
+export type LunaticGoToPage = LunaticState['goToPage']
 
-export type LunaticGetMultimode = ReturnType<typeof useLunatic>['getMultimode']
+export type LunaticGoPreviousPage = LunaticState['goPreviousPage']
+export type LunaticGoNextPage = LunaticState['goNextPage']
 
-export type LunaticGetArticulationState = ReturnType<
-  typeof getArticulationState
->
-export type LunaticArticulationStateItems = LunaticGetArticulationState['items']
+export type LunaticOverview = LunaticState['overview']
 
+export type LunaticPageTag = LunaticState['pageTag']
 
 export type Source = LunaticSource