+++ src/models/pageType.ts
@@ -1,18 +1,17 @@
 import type { LunaticState } from '@inseefr/lunatic'
 
-
 export enum PAGE_TYPE {
-    WELCOME = 'welcomePage',
-    VALIDATION = 'validationPage',
-    END = 'endPage',
-    LUNATIC = 'lunaticPage',
+  WELCOME = 'welcomePage',
+  VALIDATION = 'validationPage',
+  END = 'endPage',
+  LUNATIC = 'lunaticPage',
 }
 
-export type WalkingPaperPage =
-    | PAGE_TYPE.WELCOME
-    | PAGE_TYPE.VALIDATION
-    | PAGE_TYPE.END
+export type NirvanaPage =
+  | PAGE_TYPE.WELCOME
+  | PAGE_TYPE.VALIDATION
+  | PAGE_TYPE.END
 
-export type PageType = WalkingPaperPage | LunaticState['pageTag']
+export type PageType = NirvanaPage | LunaticState['pageTag']
 
-export type InternalPageType = WalkingPaperPage | PAGE_TYPE.LUNATIC
+export type InternalPageType = NirvanaPage | PAGE_TYPE.LUNATIC